Se cambiaron los model x los automaticos y se cambio el hibernate config

+++ Modulo/src/test/java/ar/com/magm/model/ASubgeneroTest.java
@@ -26,7 +26,7 @@ public class ASubgeneroTest extends BaseTest {
         Subgenero p = new Subgenero();
         p.setNombre("Metal");
         p = service.save(p);
-        assertNotEquals("Se generó mal el id", 0, p.getIdSubgenero());
+        assertNotEquals("Se generó mal el id", 0, (long)p.getIdSubgenero());
     }
 
     @Test
@@ -36,7 +36,7 @@ public class ASubgeneroTest extends BaseTest {
         Subgenero p = new Subgenero();
         p.setNombre("Punk");
         p = service.saveOrUpdate(p);
-        assertNotEquals("Se generó mal el id", 0, p.getIdSubgenero());
+        assertNotEquals("Se generó mal el id", 0, (long)p.getIdSubgenero());
     }
 
     @Test
@@ -52,7 +52,7 @@ public class ASubgeneroTest extends BaseTest {
         } catch (NotFoundException e) {
 
         }
-        assertNotEquals("Se generó mal el id", 0, p.getIdSubgenero());
+        assertNotEquals("Se generó mal el id", 0, (long)p.getIdSubgenero());
     }
 
     @Test
